move files to package folder
