+++ src/index.tsx
@@ -1,2 +1,3 @@
-export { useInitialMount } from "./useInitialMount";
-export { useUniqueId } from "./useUniqueID";
+export { useInitialMount } from "./package/useInitialMount";
+export { useUniqueId } from "./package/useUniqueID";
+export { usePrevious } from "./package/usePrevious";
